fix(entire): capture headless Cursor through fleet
X-Agent: codex/6291-entire-fleet-capture
